Document the goals and roadmap of the project
In the past, I've found that having specific goals for a project makes
it much easier for me to continue actively working on it in the longer
term. To facilitate this, I'm coming up with a roadmap as a list of
goals for the project beforehand, with the hope that this will help
keep me motivated.