Add assignable macros using \def
This starts building out the ability to set values via
assignment. Using the Macro struct, we build out a parsing function to
parse the body of the macro definition. Then, we set up assignment
infrastructure for parsing assignments in general, and use the macro
parser there. We also need to add a new piece of data to our TeXState
and add a setter and getter for that.

Next, we'll start using macro definitions in the expansion code.