Parse macro arguments into a parameter # -> token list map
In order to start expanding macros, we need to correctly parse the
arguments for a given macro. This is a somewhat complicated process
due to the existence of delimited parameters, where we need to keep
track of possible matches of the delimiters but still handle things
correctly when the delimiters don't end up matching.