Add very basic conditional parsing
This adds the foundation of parsing conditionals by adding support for
handling \iftrue and \iffalse. With these working, once the other
kinds of conditionals become relevant (for instance, \ifnum isn't
relevant yet because we can't set numeric variables), we can easily
add them to this framework. The next part of this will be integrating
it into the expansion code.