Adds \let assignments
This adds the ability to make assignments with \let. It turns out my
previous implementation of \let was incorrect. The TeXbook says:

> If the `<token>` is another control sequence, `\cs` will acquire the
same significance as that control sequence.
> If the `<token>` in a `\let` is a single character—i.e., if it is a
(character code, category code) pair—then the control sequence will
behave to a certain extent like that character; ...

It's not super clear about what happens if you do `\let\a=\b` when
`\b` isn't already defined to be something else. In my previous
implementation, this would set `\a` equal to the literal token `\b`,
but this is clearly incorrect. Instead, it really should just not set
`\a` to anything. This is what this new version does.

I also refactored some of the more "primitive" parsing functions that
are probably going to be used in multiple places out into a separate
module called primitives.