Add a state stack to handle grouping
When groups in TeX are entered or left with { and }, we need to keep
track of the values of states before the group started so we can
restore to those values once the group ends. We accomplish this by
keeping a stack of state objects and pushing and popping copies of the
stack when we enter and leave a group.

The other important operation that we need to be able to perform is to
apply \global assignments, which cause the assignments to apply even
after the group ends. We accomplish this by allowing assignments to be
applied to every state object in the stack. I created a new macro to
easily create these optionally-global assignment methods.