Add some comments and maybe fix \let with active tokens
I added some comments to the \let assignment code and while I was
doing that, I thought about whether \let-ting something equal to an
active token worked more like \let-ting to another kind of token or
like a control sequence. I tested it out and decided that it works
more like \let-ting to a control sequence, so I tried to fix the code
to account for that and added a test.