Allow \let-ting to primitive functions
In my previous TeX implementation, I allowed for \let-ting tokens
equal to any other token. When implementing \let here, I realized that
this was incorrect behavior, because if I \let\a=\b when \b is not
defined and then define \b, \a doesn't end up with the new
definition. Instead, we can only \let tokens equal to character tokens
or control sequences that already have definitions, so that's what
this implmentation did.

Unfortunately, there are a class of control sequences that do have
definitions but aren't macros: primitives. Before this commit, trying
to \let something to a primitive would simply have no effect.

So, this commit tries to reconcile a few of these things. I renamed
the "Let" TokenDefinition to "Token" to make it clearer that it's
mapping Token -> Token. I added a new "Primitive" TokenDefinition to
allow for \let-ting to primitives. To make that list useful, I added a
big list of all the possible primitives that this currently supports
(but will need to include new ones as they're added) and added
definitions of all of the \<primitive> control sequences to be defined
as that primitive. This way, when we \let\a=\<primitive>, we'll see
that \<primitive> is defined as a Primitive and set \a also equal to
that primitive.

However, we now need a way to check if a given token is actually equal
to that primitive. I added a new state.is_token_equal_to_cs() method
which will do a lookup to see if a token has been redefined equal to a
given control sequence. My next step will be to change the code that
is manually checking equality of control sequences to use this new
function.

Finally, it turns out that my implementation of expansion doesn't work
like it really should. If I \let\a=b, then during the expansion step
we shouldn't be turning \a into b. We know that this exchange of
control sequence for char token doesn't happen here because the
TeXbook explicitly lists things that are done during expansion, and
this step isn't listed there. Instead, I'm guessing that this
changeover happens in the list builders themselves, so I added a step
in parse_horizontal_list that decodes renamed tokens before it does
its parsing.