Check for \let-ted primitives
Now that we can \let control sequences to primitives, we need to check
for that when we're testing for those primitives in our parsing
functions. This is easy with the state.is_token_equal_to_cs() method,
which we use in this commit. I also added a bunch of tests to make
sure that this all is working.