Add count registers
This adds the 256 count registers to our state. There's no way to
interact with them yet, that will come next. Rust's primitive number
types allow us to better model the values of counts by using an i32,
which has a range that almost exactly matches a count. Unfortunately,
it's not exactly the same (the min value of a count is -2147483647,
whereas the min value of an i32 is -2147483648), so we're going to
have to do some checking anyways. Also, we're going to have other
values like dimens which contain about 31 bits of information, so
we'll have to do some checking in other places too.