Add parsing of Variable references
A lot of places in TeX, we're reading or setting different variables,
like `\count1=1` or `\box0=\hbox{x}` or `\hoffset=1in`. To aid this,
we add an abstraction around parsing variables. Next, I'll add an easy
way to set or get the value of a variable from the state. There's
currently only a single kind of variable, count variables. In the
future, I'll almost certainly need to distinguish different variables
somehow by the values that they store, but I don't need to do that
yet.

I also pulled out the `with_parser()` testing function that I built to
make it easier to create and test a parser given a set of lines.