Allow assignments to count variables
Now that we can parse variable references, we add a way to assign to
those count variables in assignments. Most of this is fairly
straightforwards.

One interesting part is how we actually do the setting. Instead of
looking at the variable in the assignment, we add some methods to
Variables so that they can set themselves given a TeXState. This is
going to become more complicated when there are variables with
different kinds of values, but I'll figure that out later.

Also, while I was at it I added a way to get the value of a (number)
variable when we parse a <number>, so we can do \count0=\count1.