Refactor is_token_equal_to_cs -> is_token_equal_to_prim
Several times, I noticed that I would start using a new primitive but
then would forget to add it to the list of primitives, which caused
problems when I used is_token_equal_to_cs. This change refactors that
function to only check for primitives (becuase that's the only thing
it was being used for anyways) and to also add an extra check during
debug mode: we ensure that every primitive that's passed in to check
is actually in the list of all primitives. This should catch cases
where I forget to add it.