Add arithmetic assignments
This adds the \advance, \multiply, and \divide assignments which
operate on integer variables. The assignments themselves are fairly
straightforwards. Things will change when there are other kinds of
variables, but with simple integer variables the assignments are very
clean. The most complicated thing here was adding optional keyword
parsing (i.e. the <optional by>).

Handling the keyword parsing raised an interesting question: If we try
to expand an active character token and there isn't a definition for
it, should we immediately bail out or should we let the parsing
function that receives the token bail out? The TeXbook says that when
parsing a keyword like pt we look for _non-active_ tokens that match
'p' and 't'. But should we even be allowing active character tokens to
come out of lex_expanded_token()? Unclear.