Allow any string-like lines to be used for lexing
Previously, only `&str`s were allowed for the lines used in
lexing. This ended up being annoying when trying to parse in the main
function, where we'll have `String`s. To make this easier, I modified
the functions which take lists of lines to accept anything
`AsRef<str>` to allow `&str`s and `String`s.