Add a useful executable for testing, see that we're turing complete
Everything is in place to try calculating primes! The final piece of
the puzzle is actually being able to parse using the executable, so
this commit adds that. We read in all of the lines from stdin and then
parse that into a horizontal list and print out the result.

Using this can be done like:

```
$ cargo run < examples/primes.tex
2, 3, 5, 7, 11, 13, 17, 19, 23, and 29
```

I was interested in whether the reimplementation in Rust fared any
better than the JS version. I changed the '10' in primes.tex to '200'
and ran:

```
$ npm run build
$ time node lib/exe.js < examples/primes.tex
2, 3, 5, 7, 11, 13, ..., 1213, 1217, and 1223

real    2.44s
user    2.64s
sys     0.12s

$ time cargo run < examples/primes.tex
2, 3, 5, 7, 11, 13, ..., 1213, 1217, and 1223

real    10.85s
user    10.76s
sys     0.05s
```

I was shocked that the rust version was more than 4 times slower!!
Fortunately, I realized that this was the unoptimized version with
debugging info enabled. I reran it on the release version:

```
$ time cargo run --release < examples/primes.tex
2, 3, 5, 7, 11, 13, ..., 1213, 1217, and 1223

real    0.64s
user    0.58s
sys     0.04s
```

Much better.