Handle nested conditionals in skipped bodies of conditionals
When we are skipping the false portion of a conditional (i.e. the `x`
in `\iffalse x\else y\fi`), we need to handle there being nested
conditionals inside of that portion. Otherwise, `\iffalse\iftrue
x\fi\fi` will end up with an extra `\fi`. This commit adds the
handling of those nested conditionals in a very simple way, by just
skipping to a corresponding \fi if we see a conditional start while
we're skipping, which I think works.