Read TFM Files
This adds structs to represent all of the different parts of a TFM
file, and a new function which uses the TeXFileReader constructs to
parse a TFM file.

I added a very basic test, but the tests are pretty onerous to write,
so I'll probably slowly add more to make me more confident of my
implementation.