Add accessors on TFMFile to get width/height/depth of a character
Getting basic dimensions of a single character is somewhat non-trivial
because of the levels of indirection in the file: the info about the
character needs to be looked up, and then the actual width needs to be
looked up based on the index. This commit adds accessors to do the
lookup.

More complicated lookups (like italic correction, kerning, ligatures,
etc.) will be added as they're needed.