Make number parsing match TeX better
The structure of how I was parsing numbers didn't match some of the
terminology in the TeXbook and also was structured incorrectly. In
particular, things like `-\count0` wouldn't have worked, because the
parsing of the `-` would be checked after we checked for register
references.

This commit fixes up the current code to match how TeX parses
numbers. I also forgot to move the number parsing tests into the
number module, so this does that while also adding a test for
`-\count0`.