Add parsing of required keywords
This adds the ability to parse keywords when they are required. This
is simpler than when parsing an optional keyword, because we don't
have to check the first character and break if it doesn't match the
keyword.

I also pulled out the "checking equal to a char ignoring case"
function to make the code a little cleaner.