Add parsing of <dimen>s
This adds the ability to parse a Dimen. This mostly follows the
grammar from the TeXbook, but deviates in a couple places where we are
unable to directly replicate it:

One place we deviate is in parsing `<factor> -> <normal integer> |
<decimal constant>`. Here, `<normal integer>` containts parse rules
for parsing an integer constant (like `123`) while `<decimal
constant>` parses a decimal (like `12.3`). If we were to try to follow
the grammar exactly when parsing `123.4`, we'd end up parsing `123` as
a `<normal integer>` but then when we'd encounter the `.` and would
know we really should be parsing a `<decimal constant>`. But we can't
always be confident that a decimal after the `<normal integer>` means
we should continue as a decimal, because `<normal integer>` also
contains internal integers, so `\count0 .4` shouldn't parse as `123.4`
if `\count0=123`. So, we manually check for some of the things in
`<normal integer>` that aren't integer constants and then handle all
integer and decimal constants separately.

The other place we deviate is in parsing units. The rules for parsing
units are broken up into "internal units" and "physical units", where
"physical units" are units that have some specific relation to real
world dimensions (like inches) whereas "internal units" have more to
do with the current state, like one em of the current font or a
current variable value. The parsing says that these are parsed
separately, but from my point of view it makes more sense to parse all
of the keyword-like units (like `em` and `pt` and `in` and `true pt`)
in one branch and handle the variable-like "units" in another branch.

Something fun I learned: You can have a variable-like in both the
factor and unit place, so something like

```
\count0=123
\count1=234
\dimen1=\count0\count1
\count2=\dimen1
```

Is a really crappy way of doing multiplication, where `\count2` ends
up equal to `\count0 * \count1`.