Allow parsing of optional keywords that don't match halfway through
Before, if we encountered the first part of a keyword but then it was
different after that, the parse would just entirely fail. This was
useful for something like the "by" in `\advance\count0 by10` because
if there was only a "b" and then we undid the parsing, then parsing
the number that follows the "by" would fail because it would
immediately see the "b".

Now, we're trying to parse the "plus" in a glue, like `1pt plus 2pt`,
where if we see part of the "plus" then we need to just finish parsing
the glue and continue with the parsing of the part of "plus".

This commit adds that ability using the `add_upcoming_tokens()` method
we use for expanding. I don't really want this to be used in more
places (because it allows us to just peek forwards as many tokens as
we want which can be bad for performance, and we really only need to
peek 1 token forwards in almost all cases), but we need it here.