Add <glue> parsing
Using a bunch of the parsing functionality I've built up in the
previous commits, parsing <glue> is now very simple. We take advantage
of the ability to parse `SpringDimen`s, which is one of the main
components of a `glue`, and we use the new functionality of optional
keyword parsing to correctly parse the "plus" and "minus" keywords.