Add \hskip parsing to horizontal lists
We add a new HorizontalListElem type for keeping track of skips. This
is fairly simple by using the new `parse_glue()` method. This commit
is mostly about cleaning up handling of primitives in horizontal_list,
deals with the new `HorizontalListElem` type, and adds a test to make
sure that \hskip parsing works.