Replace Mutex -> RefCell
Recently, I learned about `RefCell` (thanks to "Learning Rust With
Entirely Too Many Linked Lists"!), which lets us retrieve mutable
references from a non-mutable referenced object. It turns out, this
was exactly what we were using a `Mutex` for in `TeXState`, but we
aren't actually using any threading, so we can use the more
appropriate `RefCell` type to accomplish this.