Store factors of infinite length as integers
It turns out (as mentioned by the TeXbook) that the multipliers of the
different infinite lengths (i.e. fil, fill, filll) are stored as
integers, and the float versions that are written are interpreted as
multiples of 2^-16. I rewrote how these types work to account for that
and making things a little easier to deal with by adding a
FilDimen::new() function to generate..

I guessed that there was probably a maximum value for infinite lengths
(because there's maximums for everything!) and in retrospect, it's
obvious that there should be maximums for these as well. I'd be
willing to bet that *nothing* is actually stored as a floating point
number and that everything will be represented as an int because that
makes it easier to be reproducible on many machines.