Move list elements into a separate module
Something I didn't realize is that you can pull the elements that made
a box out of the box after it's been created using `\unhbox`. This
means that the implementation of boxes is going to need to store a
list of the elements that are in it. In order to accomplish this, the
definition of the list elements needs to be shared, so we have to pull
it into a separate module.