Add initial font metrics loading
As we move towards building horizontal boxes, we're going to need to
look up metrics for characters for a given font. This commit does some
of that work by adding a place to store the font metrics in our
state. Because font metric data is globally available and isn't
affected by grouping, I put this data on `TeXState` itself.

Because there isn't actually any way of loading fonts yet, I seeded
the font metrics with the font we'll be testing with: cmr10. Since I
haven't gotten kpathsea integration working just yet (but working on
it!) I had to just store the literal path to cmr10.tfm on my system.
Once the integration is figured out, I'll change that to work on other
systems.