Adds a very basic initial box creation method
This adds a very basic method of parsing horizontal boxes. In
particular, it doesn't perform "setting" the glue for
non-natural-width boxes. To accomplish this, I added a method for
having horizontal list elements return their sizes (given the current
state). Then, the list parsing combines the sizes of all of the
elements appropriately, and returns the box. Next, I'll work on adding
box glue setting.