Add cmr10.tfm file to test
In order to test a more real-world example, I added the real cmr10.tfm
metrics file to the repo so that I can test things with it. I added
some rudimentary tests from parsing the file to ensure that everything
is working fine.

I should probably check what the licensing deal is with including that
file, but for now it will be fine.