Add the ability to set glue in horizontal boxes
This adds a new parameter to box parsing for the different ways to set
the parsed box. The parameter can indicate either a FixedWidth set
(such as `\hbox to<dimen>{}`) or a Spread set (such as `\hbox
spread<dimen>{}`). The contents of the stored box aren't changed, but
we store the optional glue set ratio as a new parameter of the box.