Parse boxes via \hbox{}
This adds the ability parse a box from an `\hbox{}` command, including
`\hbox to<dimen>{}` and `\hbox spread<dimen>{}`. I also added a more
general `Box` type which wraps `HorizontalBox`es (and eventually will
also wrap `VerticalBox`es).

I had to fix a bug in `parse_horizontal_list()` where it was eating up
the trailing `}` of the list. Finally, that is important!

I also added a new primitive function for parsing `<filler>`, which is
basically just any number of spaces or `\relax`s.