Add box registers
This adds the `\box` registers used to store boxes in TeX. Becuase of
the ways that the box registers are handled (notably, when you use a
box, it becomes inaccessible afterwards both in the current stack
frame and in all of the other stack frames) the structure of how we
store boxes is different than most of the other state variables.

I also renamed `boxes::Box` to `boxes::TeXBox` becuase `Box` is
already something in common use in Rust and I don't want that to get
confusing later.