Use kpathsea library to find font metrics
Previously, I was just hardcoding a path to the cmr10.tfm file. But
now the kpathsea library is doing everything that I need it to do, so
I can use it here!

The main downside here is that I can't run the tests in parallel
because of limitations of the library. This makes the tests take ~20s
on my computer now. I might try to figure out how to reuse the kpse
instance so that I can make the tests fast again.