Run cargo tarpaulin in travis
I tried to run kcov after a successful build, but couldn't get it to
work correctly. I've tried to run tarpaulin locally, but I'm using
stable rust so it didn't work. But since I'm already running the tests
on travis with nightly rust, I can use tarpaulin! This adds that
support, and also uploads the results to coveralls.