Add DimenVariables for \wd, \ht, and \dp
This adds a new type of variable, DimenVariable, whose values are
`Dimen`s. The first kind of this variable that we're needing are
`\wd`, `\ht`, and `\dp`, which provide the width, height, and depth of
boxes in a given box register.

Something I didn't realize until testing this is that these variables
actually provide *mutable* access to the box registers! That is, you
can do `\wd0=10pt`, and that will actually change the width of the
box! I'm not sure if this is particularly useful, but it does work.

Since there was no way to get access (or in particular, mutable
access) to the boxes in the registers without pulling them out, I had
to add a new method to state, `with_box()` which takes a callback to
run on the box in that register. It actually turned out to be pretty
clean, but sadly because it has a bunch of template variables, I
couldn't use the macros to generate it (without making the macros much
more complicated).

The next step will be actually parsing these variables, and then
finding the places where dimen variables are accessible and adding
support there.