Add coerced integer dimen parsing
This adds coercing dimensions into integers when doing integer
parsing. The only thing of note here is adding the ability to get the
exact number of scaled points in a dimen, which we need for the
coercing.

Also, this brings us very close to being able to accomplish our goal!
The only thing left is to allow for putting boxes in other boxes.