Adds inner Boxes as a new horizontal element type
This adds the option to nest a TeXBox inside of a horizontal
list. There currently isn't any way for the boxes to actually get
parsed inside, that will be added next.

This raised a few concerns: for one, I was worried that this would
cause some sort of recursive type problem, but since we're storing our
elements in a list of HorizontalBox, this turns out to be fine.

The second problem was how to turn boxes into a list of chars now, for
debugging. For this, I changed the conversion from list elems into
chars from a parser function into a method on boxes. This will
generalize to vertical boxes as well.