Add parsing of \box<n> boxes
This adds the ability to parse out a box from a box register using
`\box<n>`. This removes the box from the register and returns it.

Most of the code here is self-explanatory. The most interesting thing
that happened is that `parse_box()` now has to return `Option<TeXBox>`
because a box register can be empty. This is alluded to in the TeXbook
where it says things like "if the box is void, nothing happens" or
things like that. I added some tests to make sure this is handled
correctly in the places that are using `parse_box()`.

Also, after this commit, stage 2 is done! boxes.tex correctly parses
and returns the expected output. Now all I need to do is make an
integration test to confirm that and ensure that it stays that way,
and then I can move onto stage 3!