Add a shared reference to Kpaths to make tests fast
When I added support for using Kpathsea to find the font file for
cmr10.tfm, this ended up making tests really slow for a couple
reasons. Since kpathsea isn't thread-safe, I had to run the tests
single-threaded which slowed things down. Also, since initalizing
kpathsea is really slow, each of the tests took ~500ms, which caused
the entire test run to be ~20s. This was too slow for my tastes.

To fix this, I'm now adding a shared reference to a single Kpaths
instance, guarded by a mutex. This will allow the tests to be run
multi-threaded again since there is now some locking going on, and
we're only paying the cost of initializing kpathsea once instead of
hundreds of times. Now, the tests take ~600ms total!