Add differentiation between restricted and unrestricted horizontal mode
In order to correctly handle parsing horizontal lists inside of
vertical lists, we need to know whether we're in restricted horizontal
mode or not. I accomplished this with just an argument to the parsing
function, which I think will work well enough.

The only thing that's testing for restricted mode right now is what we
do when we see `\par`: in restricted mode, this does nothing, but in
unrestricted mode it ends the list.