Move replace_renamed_token() into primitives
Since the next goal is to parse vertical lists, and that will also
require handling renamed tokens, this moves the
replace_renamed_token() function into a more general place so both of
those routines will be able to use it. Along the way, I refactored it
a bit to be more functional and added a test.