Add very basic parsing of vertical lists
This adds a foundation for parsing vertical list elements. Most of the
code here is simply doing the things that vertical lists share in
common with horizontal lists, so no new terribly interesting code was
added. I added one test to check all of the different things that this
new code supports.

Next, I'll work on actually starting to handle some cases that are
unique to vertical lists, like transitioning to horizontal lists.