Add some more special-casing for non-internal vertical mode
Since non-internal vertical mode is the mode that is used to start
(and end) parsing the document, I added some extra checking to it to
make sure that things are well-formed. In particular, we need to see a
`\end` at the end, and we refuse to end the mode for any other reason.