Make \indent and \noindent create enter horizontal mode
This is the first step towards correctly transitioning between
vertical and horizontal mode. In this commit, an explicit `\indent` or
`\noindent` is needed to trigger the transition. There is not yet any
difference between the two, and I don't do some of the steps that are
supposed to happen during that transition (like adding `\parskip` or
interline glue, or line breaking).

To make this easier, I added a new function for explicitly moving into
horizontal mode so that I don't have to deal with `BoxLayout`.

Also, I was worried that the tests that I was writing were too
long/fragile, so I'm trying something new. Instead of explicitly
writing out the entire box structure (which ends up mostly being a
bunch of copy-paste with small differences) I'm trying writing the
boxes in the input itself and then pulling the boxes out to test. That
way, if the structure of boxes changes or something, the test will
continue to function without me having to rewrite it. I'm imagining
that this will become very helpful as the boxes get larger and larger.