Make \indent add indentation to the generated hbox
This adds a new flag to `parse_horizontal_list()` for deciding whether
to add indentation to the parsed list, and then exposing that flag in
the places where we parse horizontal boxes. This allows us to
correctly generate a box with indentation when we use `\indent` in a
vertical list.