Add \copy command for retrieving copies of boxes
Now that I'm using a new way of generating the boxes for tests, having
the ability to make box copies becomes more important. This commit
adds the `\copy<num>` command for retrieving a copy of a box. I added
a test to make sure that it works, and also rewrote another test to
take advantage of this.