Handle entering horizontal mode when letters or \hskip is seen
This adds the functionality that lets the vertical list parser
automatically transition into the horizontal list parser without an
explicit `\indent` or `\noindent`. As I add new features that are used
in horizontal lists (like `\chardef`, `\unhbox`, `\vrule`, `\hfil`,
`\discretionary`, etc) I'll need to do something very similar for
those commands. I'm not sure there's a good way for me to
automatically check that I'm doing that, so I'll just have to be good
about it.