Break out of horizontal mode when vertical mode material is seen
The important thing that happens when we see vertical mode material in
horizontal mode is that `\par` is added to the input and it is allowed
to expand. I opted to make `add_upcoming_token()` public to allow for
this, instead of calling `add_upcoming_tokens()` with a single-element
vec.