Add setting glue in vertical boxes
This uses the same algorithm in `get_set_dimen_and_ratio` that's used
to set glue in horizontal lists to set the glue in vertical lists, so
this commit is pretty simple. In order to test that this is actually
working like it does in TeX, I used an example from the TeXbook to
test it.