Add parsing of \vbox
This replicates the functionality of how we parse `\hbox`es to parse
`\vbox`es, including parsing the box specification for layout. This is
pretty easy.

And with that, all of the expected work for Stage 3 is done! Running
vertical.tex works! But.. it gives incorrect output. This is due to a
few reasons:

 * Not currently adding interline glue
 * Not parsing unrestricted `\hbox` to `\hsize`.

I think I'm going to figure out how to add interline glue, and then
stop checking the resulting box width, which should be "good enough"
for making me feel confident that we're parsing vertical boxes
appropriately.