Add interline glue between boxes in vertical lists
This implements the basic interline glue algorithm. I'm using constant
values instead of the different parameters and not storing
`\prevdepth` in a parameter either to make things simpler for
now. This necessitated updating a bunch of tests, but seems to be
working correctly! Evaluating `vertical.tex` now gives the correct
(i.e. matching TeX) values for height.