Finalize Stage #3
We've gotten to the point where stage #3 goals are essentially
complete. I had to modify the end goals a tiny bit to get things
working (since I didn't want to deal with `\hsize` just quite yet),
but the important stuff is there.

At this point, I was planning on getting started with math parsing(!),
but I've decided that, to make it easier to ensure that I'm producing
the same output as real TeX, I'm going to move onto producing DVI
output. I'll probably come up with some way to compare DVI files and
see if they're content-identical (i.e. they might not have the same
exact operations, but they'll produce the same result) and start using
that to determine whether the output is correct.