Update name to XymosTeX
Originally, this project was named "rs-tex", because it was a TeX
implementation in rust. Aside from the fact that almost every other
project that is like this (i.e. something written in rust) puts the
-rs after the name instead of before, this was also not terribly
creative.

Instead, I decided to make the name of this project a pun on my
username, xymostech. I don't expect this project to become terribly
popular, so I don't think this will cause much confusion and will just
be a fun inside joke for me.