Generate fnt_def and fnt commands
One of the stateful parts of generating dvi files is keeping track of
the fonts that have been added to the file and which font is currently
selected. This needs to be kept track of so there isn't a duplicate
fnt_def and fnt command before every single character, and so that
their font numbers don't overlap. I'm doing this by storing that state
in the `DVIFileWriter`, which will be used to generate the real DVI
file.