Adds vertical boxes to dvi files
This hooks up all of the pieces to allow for adding vertical boxes to
DVI files. Since there aren't any new elements being added here, this
is pretty easy. I almost forgot to include the depth in the
calculation though, oops! Note to self: 'a' is not a great testing
character.