Add Bop/Eop commands for pages
This adds the ability to wrap top-level vertical boxes in Bop and Eop
commands. In order to do this, I need to keep track of the byte
positions of the previous pages so that I can get the previous
pointers correct. To do that, I needed to calculate the byte sizes of
each of the commands which was tedious but simple.