Make adding pages take a vertical list, not a box
Upon closer reading of how pages are added to dvi files, I see that
they aren't put into a box before being added to the page, they're
just put in as a list of vertical elems. I think that the glue set
ratio will be calculated around then so taking the glue set ratio as a
param there should work.