Add basic DVI-output skeleton
This adds a framework to do the heavy lifting of actually writing out
a DVI file on top of. I added the basic functions for writing out the
primitives (numbers, arrays, strings) that writing out the actual
commands will take. I managed to macro-ify all of the number writing,
which was nice, and added some extra tests to make sure I did it
right. (Luckily, it's the exact opposite of the reader functions, so I
was able to steal the test cases that I wrote there)