Add actual output routes for DVI commands
This adds the opposite functions of the ones in dvi/parser.rs which
parse a Reader to a DVICommand, so these take a DVICommand and output
them to a Writer. I'm also using literally the exact same test cases
that I wrote for the reader but using them here. I should probably
extract them out so that they're not just copy-pasted at some point.