Update README with new info
Mainly, I wanted to update the README because I have changed the
output of XymosTeX to be a DVI file, and the example usage didn't
account for that. However I realized that there were other things I
wanted to change about what kinds of contributions I was taking, and
also I noticed some typos in the goals so I fixed those.