Update how we enter and leave elements in vertical boxes
Previously, I wasn't taking into account how the baseline works for
vertical boxes and elements inside of vertical boxes. In particular, I
was correctly adding movement around the boxes but ended up with the
baselines for the elements inside of the boxes in the wrong spots.

This fixes that by first, pushing each element in a box down to its
baseline (based on the height) first, then rendering the element, and
then pushing the rest of element down by its width. Also, when we
enter a vertical box, we need to push back up to the top of the box so
that the baselines of the first element is correct.

With this change, the dvitest.tex file looks almost exactly like what
TeX produces, but it's not exactly lined up absolutely. I think this
has something to do with the page margins which I still haven't gotten
figured out, and might just have to do with how I'm rendering it to
pdf. I'll figure that out next.

I also found a bug in how I'm calculating the tail of the post_post,
which I'll fix in the next commit.