Add hacky handling of \topskip
The only remaining visual difference between the TeX output from the
test DVI file and my output was that everything was shifted down a
little bit in the real TeX version. After some debugging I realized
that this was due to my not handling \topskip. Really, I think that
the whole way that I am parsing non-internal vertical lists is
incorrect since they will be handled by the page builder, but for now
I added a hacky version of \topskip handling by doing something
similar to how I do \baselineskip.

Now, the visual output of TeX and XymosTeX are (as far as I can tell)
exactly the same! The next step will be writing a DVI interpreter to
confirm that they are point-for-point identical