Create DVI interpreter
This creates a new module for interpreting DVI file commands into the
positions of the resulting characters. This handles some basic
validation of the file and can interpret all of the commands in the
file that is output from interpreting the current stage goals. The
coordinates are currently interpreted in TeX points, not DVI units
(which I think are 10^-7 meters). I'll add onto it in the same way
that I add onto the dvi parser. I also added a debugging executable
for printing out the result of interpreting a file so that I can
easily compare DVI files down to the point.

The next step will be automatically testing that the output of certain
files match up with the output of TeX, which I'll use to confirm that
Stage 4 is complete. In the future, I'd also like to try doing
automated fuzzing to ensure that running TeX programs through my
version of TeX produces the same results as real TeX.