Add an integration test to ensure that the stage 4 goals are complete
This adds a test to confirm that stage 4 was actually completed. To do
this, I parse the dvitest.tex file, generate a dvi, and interpret
it. Then I go the other way and read in the expected dvi, parse it and
interpret it. I can then compare that the interpreted values are
exactly the same.

I had to make a couple tiny tweaks, like removing the final newline
from the test file and regenerating the expected dvi because
apparently it hadn't been regenerated since I last changed the test
file.

Now that this is done, I finally get to start working on my bread and
butter: math parsing!

I'll also probably write a script that takes a TeX file, runs it
through TeX and XymosTeX, and ensures that they are interpreted the
same.