Add concrete goals for Stage #5
I originally got a bit over my heels when I tried to come up with
goals for math parsing, since I've done some of this before with
KaTeX. However I decided to pull back my goals because I realized that
some of the parts of math typsetting are annoyingly hard to get right
(I'm looking at you, `\left`/`\right`). Instead, I started with a
relatively simple set of goals to be able to handle, which should be a
good test of all of the math basics.

The only "advanced" thing I'm handling is `\atop`, which I chose
because it should help me ensure that the parsing step is working
correctly without being too hard to output.