Allow assignments inside of math lists
This adds the simple support for assignments inside of math
lists. Unlike horizontal/vertical lists, where adding assignment
support is slighly more complex due to how I do the parsing, the
parsing for assignments in math lists is pretty simple, just checking
for assignments every loop. One interesting consequence of this that I
realized is that you can insert assignments in between an atom base
and its superscript/subscript, so I added a funny test that handles
this.